backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
qdrant-client>=1.7.0
google-genai>=0.4.1
langchain>=0.1.0
python-dotenv>=1.0.0
slowapi>=0.1.9
markdown>=3.5.1
pydantic>=2.5.0
typing-extensions>=4.8.0
aiofiles>=23.2.1